refactor: init prisma move ui and lib outside app
