+++ lib/definitions.ts
@@ -2,9 +2,11 @@
 // It describes the shape of the data, and what data type each property should accept.
 // For simplicity of teaching, we're manually defining these types.
 // However, these types are generated automatically if you're using an ORM such as Prisma.
-export interface User {
-  id: string;
-  name: string;
-  email: string;
-  password: string;
+export interface Product {
+  id: number;
+  title: string;
+  description: string;
+  price: number;
+  rating: number;
+  image: string;
 }
