+++ next.config.mjs
@@ -1,15 +1,15 @@
 /** @type {import('next').NextConfig} */
 const nextConfig = {
-    images: {
-        remotePatterns: [
-            {
+  images: {
+    remotePatterns: [
+      {
         protocol: 'https',
-              hostname: 'via.placeholder.com',
-              port: '',
-              pathname: '/150/**',
-            },
-          ],      
-    },
+        hostname: 'via.placeholder.com',
+        port: '',
+        pathname: '/150/**',
+      },
+    ],
+  },
 };
 
 export default nextConfig;
