+++ package.json
@@ -13,6 +13,7 @@
   },
   "dependencies": {
     "@headlessui/react": "^2.0.3",
+    "@prisma/client": "^5.14.0",
     "@vercel/postgres": "^0.8.0",
     "bcrypt": "^5.1.1",
     "next": "14.2.3",
@@ -36,6 +37,7 @@
     "lint-staged": "^15.2.4",
     "postcss": "^8",
     "prettier": "^3.2.5",
+    "prisma": "^5.14.0",
     "tailwindcss": "^3.4.1",
     "typescript": "^5"
   },
